backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.3
firebase_admin==7.3.0
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-aiplatform==1.145.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.70.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
msgpack==1.1.2
packaging==26.0
proto-plus==1.27.2
protobuf==6.33.6
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.24
requests==2.33.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
vertexai==1.43.0
websockets==16.0
